requirements.txt
pyyaml>=3.13
jinja2>=2.10
pyarrow>=0.11.1
pandas==0.24.1

dash>=0.35.2
dash-auth>=1.3.1
dash-html-components>=0.13.4
dash-core-components>=0.42.1
dash-table>=3.4.0

flask-caching>=1.4.0
flask-talisman>=0.6.0
cryptography>=2.5
